Pin DPDK packages to build environment versions
DPDK is effectively split into multiple individual packages with the RTE
version appended to the package name, for example:
"librte-pmd-nfp17.11". On some Debian/Ubuntu systems it is possible to
have access to newer versions of these packages than the ones that are
available at the time of compilation on the build system. This is
particularly true in the case of Openstack deployments where additional
repos are enabled as part of the deployment, giving the system access
to newer versions of packages.

Before this patch, when virtio-forwarder is installed, the latest
versions of the librte-* packages available on the system were marked
for installation, causing incompatibility. Now the specific versions of
these packages are added to the "Recommends" fields of the
virtio-forwarder package. When installing virtio-forwarder with a
frontend (apt/aptitude) the recommended packages will also be installed.
It is worth noting that this behavior can be modified with
"--no-recommends" and "dpkg -i <package>" will also ignore this field.

Change-Id: Ie541d7f527a4a0c3ed92fcbf1ef3055fa8efc3fc
Signed-off-by: Heinrich Kuhn <heinrich.kuhn@netronome.com>
Signed-off-by: Simon Horman <simon.horman@netronome.com>
Reviewed-by: Louis Peens <louis.peens@netronome.com>