Add meson build support
From DPDK 20.11 going forward the use of "make" has been deprecated and
replaced with meson and ninja. This commit adds the necessary meson
build files and helper scripts to build and package virtio-forwarder
using meson

This commit also adds the ability to build "static" versions of
virtio-forwarder. These static packages contain a virtio-forwarder
binary that is statically linked to the DPDK libraries.

For example, to build a static .deb: (executed in the main
virtio-forwarder repo)
    meson -Ddeb=true -Dstatic=true build
    ninja -C build build_deb

Change-Id: Ice57bd722c59d5a6448e5cc93ec2c540a3360049
Signed-off-by: Heinrich Kuhn <heinrich.kuhn@netronome.com>
Reviewed-by: Louis Peens <louis.peens@netronome.com>
Signed-off-by: Simon Horman <simon.horman@netronome.com>