Use Meson as helper to generate .sources and .dsc files
Before this patch, the Make infrastructure was used to create a
orig.tar.bz2 of the virtio-forwarder code. It is up to the user to run
the correct debuild command to generate the required files for PPA
uploads. Since the .sources and .dsc files still need to be signed by a
authorized GPG key, it is still up to the user to run debuild to sign
the files. Meson will print out the needed steps to achieve this
after this patch.

Change-Id: I14f130e68e6dc94e76cb25bac177a1d47e599f07
Signed-off-by: Heinrich Kuhn <heinrich.kuhn@netronome.com>
Reviewed-by: Louis Peens <louis.peens@netronome.com>
Signed-off-by: Simon Horman <simon.horman@netronome.com>